- Action functions now return success which allows us to properly implement failed pickups for CustomInventory. - Added CustomInventory::A_Succeed since ECWolf doesn't really have a function that can safely be used as a no-op to force Pickup success (although it's possible I'm forgetting something and yes this function is basically A_RailWait).

+++ src/g_conversation.cpp
@@ -402,7 +402,7 @@ bool ConversationModule::ParseChoiceBlock(Scanner &sc, FName key, bool isValue, 
 		case NAME_Arg3:
 		case NAME_Arg4:
 			sc.MustGetToken(TK_IntConst);
-			obj.Arg[key - (int)NAME_Arg0] = sc->number;
+			obj.Arg[(int)key - (int)NAME_Arg0] = sc->number;
 			break;
 		case NAME_NextPage:
 			sc.MustGetToken(TK_IntConst);
